Remove obsolete containerize CLI

Copilot-Session: 0489bafc-54c5-404d-b5ca-bde9210d4df0
