+++ containers.slnf
@@ -5,12 +5,10 @@
       "src\\Cli\\Microsoft.DotNet.Cli.Utils\\Microsoft.DotNet.Cli.Utils.csproj",
       "src\\Cli\\Microsoft.DotNet.InternalAbstractions\\Microsoft.DotNet.InternalAbstractions.csproj",
       "src\\Containers\\Microsoft.NET.Build.Containers\\Microsoft.NET.Build.Containers.csproj",
-      "src\\Containers\\containerize\\containerize.csproj",
       "src\\Containers\\packaging\\package.csproj",
       "test\\Microsoft.NET.Build.Containers.IntegrationTests\\Microsoft.NET.Build.Containers.IntegrationTests.csproj",
       "test\\Microsoft.NET.Build.Containers.UnitTests\\Microsoft.NET.Build.Containers.UnitTests.csproj",
-      "test\\Microsoft.NET.TestFramework\\Microsoft.NET.TestFramework.csproj",
-      "test\\containerize.UnitTests\\containerize.UnitTests.csproj"
+      "test\\Microsoft.NET.TestFramework\\Microsoft.NET.TestFramework.csproj"
     ]
   }
 }
