+++ src/Layout/redist/targets/Crossgen.targets
@@ -49,7 +49,6 @@
       <WebTools>$(InstallerOutputDirectory)Sdks\Microsoft.NET.Sdk.Web\tools\$(DefaultToolTfm)\</WebTools>
       <ProjectSystemTools>$(InstallerOutputDirectory)Sdks\Microsoft.NET.Sdk.Web.ProjectSystem\tools\$(DefaultToolTfm)\</ProjectSystemTools>
       <WorkerTools>$(InstallerOutputDirectory)Sdks\Microsoft.NET.Sdk.Worker\tools\$(DefaultToolTfm)\</WorkerTools>
-      <ContainerTasks>$(InstallerOutputDirectory)Containers\tasks\$(DefaultToolTfm)\</ContainerTasks>
     </PropertyGroup>
 
     <ItemGroup>
@@ -70,9 +69,6 @@
       <RemainingFiles Remove="$(InstallerOutputDirectory)Sdks\**\*" />
       <RemainingFiles Remove="$(InstallerOutputDirectory)**\Microsoft.TestPlatform.Extensions.EventLogCollector.dll" />
 
-      <!-- Removing Containers as containerize .DLLs shouldn't be crossgenned. -->
-      <RemainingFiles Remove="$(InstallerOutputDirectory)Containers\**\*" />
-
       <!-- Add back the .NET Core assemblies in the Sdks folder -->
       <RemainingFiles Include="$(NetSdkAnalyzers)**\*" />
       <RemainingFiles Include="$(NetSdkTools)**\*" />
@@ -84,7 +80,6 @@
       <RemainingFiles Include="$(WebTools)**\*" />
       <RemainingFiles Include="$(ProjectSystemTools)**\*" />
       <RemainingFiles Include="$(WorkerTools)**\*" />
-      <RemainingFiles Include="$(ContainerTasks)**\*" />
 
       <!-- Don't crossgen VB since the usage doesn't justify the size cost for everyone -->
       <RoslynFiles Remove="$(InstallerOutputDirectory)**\Microsoft.CodeAnalysis.VisualBasic*.dll" />
