+++ src/Layout/redist/targets/GenerateLayout.targets
@@ -296,11 +296,6 @@
       Targets="Publish"
       Projects="$(RepoRoot)/src/Containers/Microsoft.NET.Build.Containers/Microsoft.NET.Build.Containers.csproj"
       Properties="Configuration=$(Configuration);PublishDir=$(OutputPath)/Containers/tasks/$(SdkTargetFramework);TargetFramework=$(SdkTargetFramework)" />
-    <MSBuild
-      Targets="Publish"
-      Projects="$(RepoRoot)/src/Containers/containerize/containerize.csproj"
-      Properties="Configuration=$(Configuration);PublishDir=$(OutputPath)/Containers/containerize"
-      Condition="'$(OSName)' == 'win'" />
   </Target>
 
   <Target Name="GenerateCliRuntimeConfigurationFiles" DependsOnTargets="PublishTargetExtensions">
