--- test/containerize.UnitTests/containerize.UnitTests.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="MSTest.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>$(SdkTargetFramework)</TargetFrameworks>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <IsTestProject>true</IsTestProject>
-    <StrongNameKeyId>MicrosoftShared</StrongNameKeyId>
-    <MSTestParallelizeScope>MethodLevel</MSTestParallelizeScope>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Containers\containerize\containerize.csproj" />
-  </ItemGroup>
-
-</Project>
